chore: split the gist-source test by behaviour
- What: `tests/http/menu/gist-source.test.ts` 393 → **200**, with override mode,
  reset-all and the official-source endpoint moving to
  `gist-source-override.test.ts` (238). Allow-list entry removed.

- Why: #2833, and CLAUDE.md's "nested, one behavior per file" — six describe
  blocks in one file.

- Impact: none. Same 125 tests pass in `tests/http/menu/` as before.

Dead imports pruned from both halves rather than copied along: `db`/`menuItems`/
`eq` are only used by the override suite, `getMenuConfig`/`getSetting` only by the
source-reading one.

The override half keeps the fetch and env restores, and its header says why:
`--isolate` gives each FILE a fresh global, not each test, so a suite that replaces
`globalThis.fetch` or mutates `process.env` still has to put them back between tests.

Gate: `bunx tsc --noEmit` exit 0 · `bun test --isolate tests/http/menu/ tests/build/`
196 pass.

Allow-list: 22 at #2833's filing → **14**.

Refs #2833
